long double

+++ LPsolver/LPsolver/main.cpp
@@ -24,7 +24,7 @@ int main(int argc, const char * argv[])
 
     int advertiser_dimensions [] = {1000000};
     int impression_dimensions [] = {1000000};
-    double bid_sparsity_scenario [] = {0.0001};
+    long double bid_sparsity_scenario [] = {0.0001};
 
     std::string file_name_path = "/Users/ciocan/Documents/Google/data/experiment_";
     int num_iterations = 100;
